feat: Implemente o RestaurantController com endpoints REST Post e Get.
