+++ src/main/java/com/ordemflow/ordemFlow/bussiness/DTOs/RestaurantResponse.java
@@ -1,4 +1,4 @@
-package com.ordemflow.ordemFlow.bussines.DTOs;
+package com.ordemflow.ordemFlow.bussiness.DTOs;
 
 import com.ordemflow.ordemFlow.infrastructure.Entity.Address;
 import lombok.Builder;
@@ -8,6 +8,6 @@ import java.time.LocalDateTime;
 
 @Builder
 public record RestaurantResponse(Long id, String name, String description,
-                                 String phone, String email,String imagemUrl, Address address,
+                                 String phone,String email,String imagemUrl, Address address,
                                  boolean active, LocalDateTime createdAt) {
 }
